feat : question-detail 질문방 생성 및 질문 생성

+++ src/apis/questions/questions.service.ts
@@ -1,16 +1,22 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Question } from './entities/question.entity';
-import { CreateQuestionDto } from './dto/create-question.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ChatgptsService } from '../chatgpts/chatgpts.service';
 import {
+  IQuestionServiceContinueQuestion,
   IQuestionServiceDeleteQuestion,
   IQuestionServiceFindAllQuestion,
   IQuestionServiceFindQuestion,
   IQuestionsServiceCreateQuestion,
 } from './interfaces/questions-service.interface';
 import { QuestionDetailsService } from '../questionDetails/question-details.service';
+import { MessageResDto } from 'src/commons/dto/message-res.dto';
 
 @Injectable()
 export class QuestionsService {
@@ -21,55 +27,40 @@ export class QuestionsService {
     private readonly questionDetailsService: QuestionDetailsService,
   ) {}
 
-  // 질문 계속
-  async continueQuestion({ questionId, userId, createQuestionDto }) {
-    const { query } = createQuestionDto;
-    const question = await this.findQuestion({ questionId, userId });
+  // 질문하기
+  async continueQuestion({
+    questionId,
+    userId,
+    continueQuestionDto,
+  }: IQuestionServiceContinueQuestion): Promise<string> {
+    const { query } = continueQuestionDto;
 
-    if (!question) {
-      throw new HttpException(
-        '질문을 찾을 수 없습니다.',
-        HttpStatus.BAD_REQUEST,
-      );
-    }
+    const question = await this.findQuestion({ questionId, userId });
 
     const questionDetails =
       await this.questionDetailsService.findQuestionDetail(questionId);
 
+    if (questionDetails.length > 20) {
+      throw new ForbiddenException('최대 질문 개수가 초과되었습니다.');
+    }
     const chatgptAnswer = await this.chatgptsService.createChatgpt(
       questionDetails,
-      createQuestionDto,
+      continueQuestionDto,
     );
 
-    await this.questionDetailsService.createQuestionDetail(
+    await this.questionDetailsService.createQuestionDetail({
       questionId,
       query,
       chatgptAnswer,
-    );
+    });
 
     return chatgptAnswer;
   }
 
-  // 질문 생성
-  async createQuestion({
-    userId,
-    createQuestionDto,
-  }: IQuestionsServiceCreateQuestion): Promise<Question> {
-    // 질문가져오기
-    const question = createQuestionDto.title;
-
-    // 질문을 챗 gpt에게 넘겨서 답변 받기
-    // const chatgptAnswer = await this.chatgptsService.createChatgpt(question);
-
-    // 답변을 JSON으로 변환
-    // const responseData = JSON.parse(chatgptAnswer);
-
-    // responseData 메세지를 createQuestionDto.answer에 삽입
-    // createQuestionDto.answer = responseData.message;
-
+  // 질문 방 생성
+  async createQuestion({ userId }): Promise<Question> {
     return await this.questionsRepository.save({
       user: { id: userId },
-      ...createQuestionDto,
     });
   }
 
@@ -80,7 +71,7 @@ export class QuestionsService {
     const { page, size } = pageReqDto;
 
     const questions = await this.questionsRepository.find({
-      order: { createdAt: 'ASC' }, // 최근순이 아래로 나오게
+      order: { createdAt: 'DESC' },
       take: size,
       skip: (page - 1) * size,
     });
@@ -95,6 +86,7 @@ export class QuestionsService {
     const questions = await this.questionsRepository.find({
       order: { createdAt: 'DESC' },
       take: count,
+      where: { user: { id: userId } },
     });
 
     return questions;
@@ -107,6 +99,7 @@ export class QuestionsService {
   }: IQuestionServiceFindQuestion): Promise<Question> {
     const question = await this.questionsRepository.findOne({
       where: { id: questionId },
+      relations: ['user', 'questionDetails'],
     });
 
     if (!question) {
@@ -114,6 +107,10 @@ export class QuestionsService {
         '질문을 찾을 수 없습니다.',
         HttpStatus.BAD_REQUEST,
       );
+    }
+
+    if (question.user.id !== userId) {
+      throw new ForbiddenException('권한이 없습니다.');
     }
 
     return question;
@@ -123,9 +120,10 @@ export class QuestionsService {
   async deleteQuestion({
     userId,
     questionId,
-  }: IQuestionServiceDeleteQuestion): Promise<void> {
+  }: IQuestionServiceDeleteQuestion): Promise<MessageResDto> {
     const question = await this.questionsRepository.findOne({
       where: { id: questionId },
+      relations: ['user'],
     });
 
     if (!question) {
@@ -135,6 +133,11 @@ export class QuestionsService {
       );
     }
 
+    if (question.user.id !== userId) {
+      throw new ForbiddenException('권한이 없습니다.');
+    }
+
     await this.questionsRepository.remove(question);
+    return { message: '질문이 삭제되었습니다.' };
   }
 }
